# modules/openapi/api.yaml
swagger: '2.0'
info:
  title: GeoGig Web API
  version: "0.2.0"
  description: 'GeoGig Web API.  You can find out more about
    GeoGig at [http://geogig.org](http://geogig.org)'
  termsOfService: http://geogig.io/terms/
  contact:
    email: groldan@boundlessgeo.com
  license:
    name: Eclipse Distribution License - v 1.0
    url: https://github.com/locationtech/geogig/blob/master/LICENSE.txt
# host: localhost:8182
# basePath: /
produces: 
- 'application/json'
- 'application/x-jackson-smile'
...
schemes:
 - http
 - https
securityDefinitions:
  BasicAuth:
    type: basic
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-GeoGig-API-Key
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: https://example.com/oauth/authorize
    tokenUrl: https://example.com/oauth/token
    scopes:
      read: Grants read access
      write: Grants write access
      admin: Grants read and write access to administrative information
security: #This makes all end points work with any of the three auth mechanisms... so far so good.
  - BasicAuth: []
  - ApiKeyAuth: []
  - OAuth2: []
parameters:
...
  userName: 
    name: user
    in: path
    required: true
    description: repository login owner
    type: string
...
  repoName: 
    name: repo
    in: path
    required: true
    description: repository name
    type: string
...
  layerName: 
    name: layer
    in: path
    required: true
    description: layer name for feature service
    type: string
...
  head: 
    name: head
    in: header
    required: false
    description: a refSpec string leading to a root RevTree
    type: string
    default: WORK_HEAD
    minItems: 1
    maxItems: 1
...
  transactionOptional:
    name: geogig-transaction-id
    in: header
    required: false
    description: Optional transaction identifier
    type: string
    format: uuid
...
paths:
...
  /layers/{user}:
    parameters: #parameters common to all methods
      - $ref: '#/parameters/userName'
    get:
      tags: 
      - Feature Service
      summary: 'return a summary of this user''s layers'
      operationId: getUserLayersSummaries
      responses:
        200:
          description: 'An array of RepoLayerSummary objects'
          schema:
            type: array
            items:
              $ref: '#/definitions/RepoLayerSummary'
  /layers/{user}/{repo}:
    parameters: #common parameters to all methods
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
    get:
      parameters: 
      - $ref: '#/parameters/transactionOptional'
      - $ref: '#/parameters/head'
      tags:
      - Feature Service 
      summary: 'List layers in a repo'
      description: 'List the available layers in the requested [transaction]/repository/[head]'
      operationId: listLayers
      responses:
        200:
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/LayerInfo'
        404:
          description: Object does not exist
    post:
      tags: 
      - Feature Service
      summary: 'Create a new layer'
      description: 'Creating and populating a new layer are two separate steps. First the layer schema must be created and then the features uploaded.'
      operationId: createLayer
      parameters:
      - $ref: '#/parameters/transactionRequired'
      - $ref: '#/parameters/featureType'
      responses:
        201:
          description: 'Layer created'
          schema:
            $ref: '#/definitions/LayerInfo'
        401:
          description: 'Unauthorized. Authentication required.'
        403:
          description: 'Forbidden. Authenticated user has no permissions to create a layer on this repository.'
        404:
          description: 'Repository or transaction state not found'
  /layers/{user}/{repo}/{layer}:
    parameters: #parameters common to all methods
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    head:
      tags:
      - Feature Service
      summary: 'Get a unique identifier for the current state of the layer.'
      description: 'Get a unique identifier for the current state of the layer. Any change/edit to the layer will change this identifier. This operation can be used to verify whether the layer has changed since the client last accessed it.'
      operationId: getLayerHash
      parameters:
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      produces:
      - application/json
      - text/plain
      responses:
        200:
          description: 'Return the current hash identifier for the layer'
          schema:
            $ref: '#/definitions/ObjectHash'
    get:
      tags:
      - Feature Service
      summary: 'Get layer metadata at the specified transaction/head'
      operationId: getLayerInfo
      parameters:
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      - name: If-None-Match
        type: string
        in: header
        description: 'Conditional get, layer info if its revision tree hash does not match the provided hash'
      responses:
        200:
          description: 'successful operation, layer metadata for the requested [transaction]/layer/[head]'
          schema:
            $ref: '#/definitions/LayerInfo'
          headers:
            ETag:
              type: string
              description: 'Hash of the current layer tree at the specified transaction/head. Can be used as the If-None-Match header for the next request to make a conditional request with the same parameters.'
        304:
          description: 'Not modified, returning no content. Layer tree hash matches the If-None-Match header parameter.'
        404:
          description: 'Layer not found'
    delete:
      tags: 
      - Feature Service
      summary: 'Delete a Layer from a repository at the current transaction''s branch'
      operationId: deleteLayer
      parameters:
      - $ref: '#/parameters/transactionRequired'
      responses:
        204:
          description: The server successfully deleted the Layer and is not returning any content
        403:
          description: The Authenticated user has no rights to delete the given Layer
        404:
          description: Layer not found
  /layers/{user}/{repo}/{layer}/bounds:
    parameters: #parameters common to all methods
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    get:
      tags:
      - Feature Service
      summary: 'Get layer bounds at the specified transaction/head. If no head is specified, defaults to WORK_HEAD'
      operationId: getBounds
      parameters:
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      responses:
        200:
          description: 'successful operation, layer bounds for the requested [transaction]/layer/[head]'
          schema:
            $ref: '#/definitions/BoundingBox'
        404:
          description: 'Layer not found'
  /layers/{user}/{repo}/{layer}/size:
    parameters: #parameters common to all methods
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    get:
      tags:
      - Feature Service
      summary: 'Get layer''s number of features at the specified transaction/head. If no head is specified, defaults to WORK_HEAD'
      operationId: getSize
      parameters:
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      responses:
        200:
          description: 'successful operation, layer size for the requested [transaction]/layer/[head]'
          schema:
            type: integer
            format: int64
        404:
          description: 'Layer not found'
  /layers/{user}/{repo}/{layer}/schema:
    parameters: #parameters common to all methods
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    get:
      tags:
      - Feature Service
      summary: 'Get Layer schema'
      description: 'Get layer''s Feature Type at the specified transaction/head. If no head is specified, defaults to WORK_HEAD'
      operationId: getSchema
      parameters:
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      responses:
        200:
          description: 'successful operation, layer size for the requested [transaction]/layer/[head]'
          schema:
            $ref: '#/definitions/RevisionFeatureType' 
        404:
          description: 'Layer not found'
    put:
      tags:
      - Feature Service
      summary: 'Change Layer schema. Must be called inside a transaction.'
      description: 'Modify the layer''s number of features at the specified transaction/head. If no head is specified, defaults to WORK_HEAD'
      operationId: updateSchema
      consumes:
      - application/json
      parameters:
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionRequired'
      - $ref: '#/parameters/featureType'
      responses:
        200:
          description: 'successful operation, the updated schema. May differ slightly from the provided one to match layer name'
          schema:
            $ref: '#/definitions/RevisionFeatureType' 
        404:
          description: 'Layer not found'
  /layers/{user}/{repo}/{layer}/tiles/{z}/{x}/{y}.mvt:
    parameters: 
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    get:
      tags:
      - Feature Service
      summary: Get a Mapbox Vector Tile
      operationId: getTile
      parameters:
      - name: z
        in: path
        required: true
        type: integer
        format: int32
      - name: x
        in: path
        required: true
        type: integer
        format: int32
      - name: y
        in: path
        required: true
        type: integer
        format: int32
      - $ref: '#/parameters/attributes'
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      - name: If-None-Match
        type: string
        in: header
        description: 'Conditional get, return features if the layer tree does not match the provided hash'
      produces:
      - application/vnd.mapbox-vector-tile
      responses:
        200:
          description: 'successful operation, a MapBox vector tile'
          schema:
            type: string
            format: binary
          headers:
            ETag:
              type: string
              description: 'Hash of the current layer tree at the specified transaction/head. Can be used as the If-None-Match header for the next request to make a conditional request with the same parameters.'
        304:
          description: 'Not modified, returning no content. Layer tree hash matches the If-None-Match header parameter.'
        404:
          description: 'Layer not found, or the requested tile does not intersect the layer bounds' 
  /layers/{user}/{repo}/{layer}/{revA}/{revB}/difftiles/{z}/{x}/{y}.mvt:
    parameters: 
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    get:
      tags:
      - Feature Service
      summary: Get a Mapbox Vector Tile
      operationId: getDiffTile
      parameters:
      - name: z
        in: path
        required: true
        type: integer
        format: int32
      - name: x
        in: path
        required: true
        type: integer
        format: int32
      - name: y
        in: path
        required: true
        type: integer
        format: int32
      - name: revA
        in: path
        required: true
        type: string
      - name: revB
        in: path
        required: true
        type: string
      - $ref: '#/parameters/attributes'
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      - name: If-None-Match
        type: string
        in: header
        description: 'Conditional get, return features if the layer tree does not match the provided hash'
      produces:
      - application/vnd.mapbox-vector-tile
      responses:
        200:
          description: 'successful operation, a MapBox vector tile'
          schema:
            type: string
            format: binary
          headers:
            ETag:
              type: string
              description: 'Hash of the current layer tree at the specified transaction/head. Can be used as the If-None-Match header for the next request to make a conditional request with the same parameters.'
        304:
          description: 'Not modified, returning no content. Layer tree hash matches the If-None-Match header parameter.'
        404:
          description: 'Layer not found, or the requested tile does not intersect the layer bounds'          
  /layers/{user}/{repo}/{layer}/features:
    parameters: 
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    get:
      tags:
      - Feature Service
      summary: Get Features from a geogig Layer
      description: 'Obtain a stream of features from a geogig layer at the specified transaction/head'
      operationId: getFeatures
      parameters:
      - $ref: '#/parameters/bbox'
      - $ref: '#/parameters/attributes'
      - $ref: '#/parameters/head'
      - $ref: '#/parameters/transactionOptional'
      - $ref: '#/parameters/page'
      - $ref: '#/parameters/pageSize'
      - name: If-None-Match
        type: string
        in: header
        description: 'Conditional get, return features if the layer tree does not match the provided hash'
      produces:
      - application/vnd.geo+json  #GeoJSON
      - application/vnd.geo+smile #binary encoded geojson
      - application/vnd.geogig.features+json #simplified geojson
      - application/vnd.geogig.features+smile #binary simplified geojson
      - application/geogig.x-protobuf # proto3
      responses:
        200:
          description: 'successful operation, stream of features in the requested [transaction]/layer/[head]'
          schema:
            $ref: '#/definitions/FeatureCollection'
          headers:
            ETag:
              type: string
              description: 'Hash of the current layer tree at the specified transaction/head. Can be used as the If-None-Match header for the next request to make a conditional request with the same parameters.'
        304:
          description: 'Not modified, returning no content. Layer tree hash matches the If-None-Match header parameter.'
        404:
          description: 'Layer not found'
    post:
      tags:
      - Feature Service
      summary: Iserts features to this collection in the WORK_HEAD. Parameter features can be either a Feature or a FeatureCollection.
      description: The features are added to the WORK_HEAD tree. This operation works against the currently checked out branch for the indicated transaction, and does not perform a geogig commit on its own. The geogig commit is created either manually or once the geogig transaction is committed.
      operationId: addFeatures
      consumes:
      - application/vnd.geo+json
      - application/vnd.geo+smile
      - application/geogig.x-protobuf
      parameters:
      - name: features
        in: body
        schema:
          $ref: '#/definitions/Feature'
      - $ref: '#/parameters/transactionRequired'
      responses:
        204:
          description: 'successful operation'
        404:
          description: 'Layer not found'
    delete:
      tags:
      - Feature Service
      summary: Deletes all features in the layer
      operationId: truncate
      parameters:
      - $ref: '#/parameters/transactionRequired'
      responses:
        204:
          description: 'successful operation'
        404:
          description: 'Layer not found'
  /layers/{user}/{repo}/{layer}/rpc/query:
    parameters: 
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    post:
      tags:
      - Feature Service
      summary: Query and return features of this layer tha match the specified FeatureQuery 
      description: The features are looked up WORK_HEAD tree unless the head parameter has been provided. 
      operationId: queryFeatures
      parameters:
      - name: query
        in: body
        required: false
        schema:
          $ref: '#/definitions/FeatureQuery'
      - $ref: '#/parameters/transactionOptional'
      produces:
      - application/vnd.geo+json  #GeoJSON
      - application/vnd.geo+smile #binary encoded geojson
      - application/vnd.geogig.features+json #simplified geojson
      - application/vnd.geogig.features+smile #binary simplified geojson
      - application/geogig.x-protobuf # proto3
      responses:
        200:
          description: 'successful operation, stream of features in the requested [transaction]/layer/[head]'
          schema:
            $ref: '#/definitions/FeatureCollection'
          headers:
            ETag:
              type: string
              description: 'Hash of the current layer tree at the specified transaction/head. Can be used as the If-None-Match header for the next request to make a conditional request with the same parameters.'
        404:
          description: 'Layer not found'
  /layers/{user}/{repo}/{layer}/rpc/update:
    parameters: 
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    post:
      tags:
      - Feature Service
      summary: Modifies features matching the given query with the given values, in the WORK_HEAD of the currently checked out branch for the indicated transaction
      description: The features are looked up and modifies in current  WORK_HEAD tree. This operation works against the currently checked out branch for the indicated transaction, and does not perform a geogig commit on its own. The geogig commit is created either manually or once the geogig transaction is committed.
      operationId: modifyFeatures
      parameters:
      - name: query
        in: body
        schema:
          $ref: '#/definitions/UpdateFeaturesRequest'
      - $ref: '#/parameters/transactionRequired'
      responses:
        204:
          description: 'successful operation'
        404:
          description: 'Layer not found'
  /layers/{user}/{repo}/{layer}/rpc/delete:
    parameters: 
      - $ref: '#/parameters/userName'
      - $ref: '#/parameters/repoName'
      - $ref: '#/parameters/layerName'
    post:
      tags:
      - Feature Service
      summary: Delete features matching the given query
      operationId: deleteFeatures
      parameters:
      - name: filter
        in: body
        schema:
          $ref: '#/definitions/FeatureFilter'
      - $ref: '#/parameters/transactionRequired'
      responses:
        204:
          description: 'successful operation'
        404:
          description: 'Layer not found'
...
definitions:
...
  LayerInfo:
    type: object
    description: 'Single Layer metadata at a given head/branch'
    properties:
      head:
        type: string
        description: 'name of the branch or other head ref-spec (commit id, tag name, etc) where the layer info exists as advertised'
      name:
        type: string
      title:
        type: string
      abstract:
        type: string
      size:
        type: integer
        format: int64
        default: 0
        minimum: 0
      bounds:
        $ref: '#/definitions/BoundingBox'
      type:
        $ref: '#/definitions/RevisionFeatureType'
...
  RepoLayerSummary:
    type: object
    description: 'Summary of layers in a repository'
    properties:
      repository:
        type: string
      layers:
        type: array
        items:
          $ref: '#/definitions/LayerSummary'
...